// src/app/shared/models/redisLedgerEntry.ts
import { Guid } from '../functions/uuid';

export interface IBasket {
  id: string;
  items: IBasketItem[];
  description: string;
  entryDate: Date;
  cubeControl: number;
}

export interface IBasketItem {
  id: string;
  dcOption: string;
  amount: number;
  account: string;
  tAccount: string;
}

export class Basket implements IBasket {
  id = Guid();
  items: IBasketItem[] = [];
  description: string;
  entryDate: Date;
  cubeControl: number;
}